sched: event throttling bugfix and cleanup
When both time and event counter throttling enabled, throttled
tasks were re-enabled due to the bug in slack timer code that is
intended to distributed remaining slack time to other cores (if any).
Due to the complexity of handling both time and events, currently
slack timer is disabled when both time and event throttling are enbled
for the sake of simplicity.