sched: add coreidle mechanism
The coreidle mechanism is a simple way to idling specified cores.

It uses a cpumap, sched_coreidle_mask, to indiciate which cores
to be idle. Once a cpu is set on the sched_coreidle_mask, the core
will be excluded from future task scheduling (either migration or
wakeup). However, it does not immediately migrate tasks that currently
run on the core. Those tasks will be migrated only on the next
scheduling point. Therefore there's almost no overhead of using
this mechanism.

The mechanism is intended to be used by kernel scheduler itself or
userspace scheduler which want to coordinate with the kernel scheduler.

You can think it as a light-weight version of hotplug mechanism.