color: bring back expand_index and bug fix
Fix hang problem in rmqueue_smallest() when used with index.
The problem arises when the requesting 'order' is non-zero.
Suppose, for example, the requesting order is 1 (size=2^1=2) and
we have found a matching page in a list of order 2 as shown below.
Now, the index of the matching page is 1.

0 1 2 3
  ^
index=1
order=1
current_order=2
return page 1

Although expand_index() will return page [2 3] and page [0]
correctly to the corresponding freelists (order 1 and order 0
respectively), the returning page is assumed to be the size of 2
(i.e., page [1 2]). Therefore, the page 2 will be overwritten
although it is in the order 1 freelist.

In order to maintain consistency, using index should be limited
only when the requesting order is zero.